refactor: change storage name function
