+++ packages/analytics-core/src/storage/storage-name.ts
